Rework Input/ Fixed Hole / Countdown on start
Reworked the input to be more consistent with direction, magnitude of swipe and speed of swipe motion. Made the hole bigger and modified the faces of the board to be one face - this has seemed to fix the issue with the hole being an obstacle at times. Added physics material to board that has very slight bounce, some dynamic friction and small static friction. The slime collider has a frictionless material.